requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cached-property==1.5.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.100.0
fitfile==1.1.9
garmindb==3.6.3
garth==0.5.2
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idbutils==1.1.0
idna==3.10
jiter==0.9.0
llm==0.23
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.3
oauthlib==3.2.2
openai==1.69.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==6.30.2
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pylance==0.23.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
markdown==3.8.2
python-ulid==3.0.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
starlette==0.27.0
tabulate==0.9.0
tcxfile==1.0.4
tornado==6.4.2
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
youtube-transcript-api==1.2.4
